+++ app/assets/stylesheets/dashboard.scss
@@ -4,6 +4,6 @@
 button {
   &:hover .t-hover-white {
     fill: white;
-    transition: 200ms;
+    transition: 100ms;
   }
 }
