+++ app/controllers/dashboard_controller.rb
@@ -1,5 +1,6 @@
 class DashboardController < ApplicationController
   def index
     @workflows = Workflow.limit(20)
+    @workflow = false
   end
 end
