+++ app/controllers/workflows_controller.rb
@@ -1,12 +1,13 @@
 class WorkflowsController < ApplicationController
 
   def create
-    new_workflow = Workflow.new(users_id: current_user.id, title: 'New Workflow')
+    new_workflow = Workflow.new(title: 'New Workflow')
+    new_workflow.creator = current_user
     new_workflow.save
-    @workflows = Workflow.limit(20)
 
     respond_to do |format|
-      format.text { render 'shared/workflows_list', workflows: @workflows, formats: [:html] }
+      format.html { render 'dashboard/index' }
+      format.text { render partial: 'shared/workflow', locals: { workflow: new_workflow } }
     end
   end
