--- app/javascript/controllers/workflow_controller.js
@@ -1,10 +1,0 @@
-import { Controller } from "stimulus"
-
-export default class extends Controller {
-
-  static targets = ["activated"]
-
-  connect() {
-    console.log(this.activatedTarget);
-  }
-}
