+++ app/views/dashboard/index.html.erb
@@ -1,4 +1,4 @@
-<div class="flex justify-between h-screen">
+<div data-controller="world" class="flex justify-between h-screen">
 
   <div class="w-72 h-screen p-5 inline-block shadow-lg">
     <h1 class="font-medium pb-3">🅗Taskete</h1>
