+++ app/views/shared/_content_main.html.erb
@@ -1,12 +1,16 @@
-<div class="inline-block rounded-xl bg-purple-100 py-2 px-4 mb-3">
-  <h1 class="text-sm">Title workflow</h1>
-</div>
+<div data-world-target="contentmain">
+  <div class="h-10"></div>
 
-<div class="mb-5 h-screen rounded-lg border mx-auto relative flex">
-  <div class="flex-grow">
-  </div>
-  <div class="w-10 py-5 h-full border-l flex flex-col items-center">
-    <div class="rounded-full w-3 h-3 border- mb-2"></div>
-    <div class="rounded-full w-3 h-3 bg-purple-400 mb-2"></div>
+  <div class="mb-5 h-screen rounded-lg border mx-auto relative flex">
+
+    <%# Content here %>
+    <div class="flex-grow">
+
+    </div>
+
+    <div class="w-10 py-5 h-full border-l flex flex-col items-center">
+      <div class="rounded-full w-3 h-3 bg-purple-400 mb-2"></div>
+      <div class="rounded-full w-3 h-3 bg-purple-400 mb-2"></div>
+    </div>
   </div>
 </div>
