+++ app/views/shared/_nav_account.html.erb
@@ -8,6 +8,9 @@
           <div class="bg-purple-400 border-2 rounded-full h-8 w-8 flex items-center justify-center text-white text-sm">
             <%= current_user.first_name.first.capitalize %><%= current_user.last_name.first.capitalize %>
           </div>
+          <div class="text-sm">
+            <%= current_user.first_name.capitalize %>&nbsp;<%= current_user.last_name.capitalize %>
+          </div>
         </div>
 
         <div class="px-2">
