+++ app/views/shared/_sidebar_l.html.erb
@@ -1,7 +1,7 @@
 <div data-controller="sidebar-l">
 
   <%# create button %>
-  <button data-action="click->sidebar-l#createBtnDropdown" class="w-full border border-purple-400 rounded-lg p-2 mt-2 flex items-center hover:bg-purple-400 hover:text-white duration-200">
+  <button data-action="click->sidebar-l#createBtnDropdown" class="w-full border border-purple-400 rounded-lg p-2 flex items-center hover:bg-purple-400 hover:text-white duration-100">
     <div class="px-2">
       <svg class="t-hover-white" xmlns="http://www.w3.org/2000/svg" width="10" height="10" viewBox="0 0 10 10">
         <path id="mini_plus" d="M44,743H41v-3a1,1,0,0,0-2,0v3H36a1,1,0,0,0,0,2h3v3a1,1,0,0,0,2,0v-3h3a1,1,0,0,0,0-2Z" transform="translate(-35 -739)"/>
@@ -12,22 +12,22 @@
 
   <%# create button dropdown %>
   <div data-sidebar-l-target="dropdown" class="hidden py-2 mt-1 w-48 bg-white rounded-md shadow-xl fixed hidden">
-    <button data-action="click->sidebar-l#createWorkflow" class="block w-full flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">
+    <button data-action="click->world#createWorkflow" class="block w-full flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">
       Workflow
     </button>
   </div>
 
   <%# tab %>
-  <div class="flex gap-4 mt-6">
+  <div class="flex gap-4 my-5">
     <button class="tablinks text-sm font-medium text-purple-600 pb-1 border-b-2 border-purple-600" >My Workflows</button>
     <button class="tablinks text-sm font-medium text-gray-400 pb-1 border-b-2 border-white" >My Tasks</button>
   </div>
 
   <%# search bar form to search within tab %>
-  <form action="" class="flex gap-4 justify-center bg-white border-b py-2">
+  <form action="" class="flex gap-2 justify-center bg-white border-b py-2">
     <button type="submit">
       <div class="">
-        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20">
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 20 20">
           <path id="ic24-search" d="M11,2a9,9,0,0,1,7.032,14.617l3.675,3.676a1,1,0,0,1-1.32,1.5l-.094-.083-3.676-3.675A9,9,0,1,1,11,2Zm0,2a7,7,0,1,0,7,7A7,7,0,0,0,11,4Z" transform="translate(-2 -2)" fill-rule="evenodd"/>
         </svg>
       </div>
@@ -37,9 +37,8 @@
   </form>
 
   <%# workflows list %>
-  <div data-sidebar-l-target="workflowslist" class="px-2 mt-4">
+  <div data-world-target="workflowslist" class="mt-4">
     <%= render 'shared/workflows_list', workflows: @workflows %>
   </div>
 
-  <%= link_to 'create', workflows_path, method: 'post' %>
 </div>
