+++ app/views/shared/_workflow.text.erb
@@ -1,0 +1,3 @@
+<div data-id="<%= workflow.id %>" data-world-target="workflow" data-action="click->sidebar-l#selectWorkflow" class="px-4 py-2 my-1 text-sm rounded-lg bg-purple-100 duration-100">
+  <%= workflow.title %>
+</div>
