+++ app/views/shared/_workflow_expand.text.erb
@@ -1,0 +1,3 @@
+<div class="inline-block rounded-xl bg-purple-100 py-2 px-4 mb-3">
+  <h1 class="text-sm"><%= @workflow.title.capitalize %></h1>
+</div>
